@@ -13,9 +13,8 @@ class ForwardSubmitButton extends StatelessWidget {
       padding: const EdgeInsets.symmetric(horizontal: 8.0),
       child: ElevatedButton(
         onPressed: onPressed,
-        style: ButtonStyle(
-          fixedSize: MaterialStateProperty.all(
-              Size(MediaQuery.of(context).size.width, 50)),
+        style: ElevatedButton.styleFrom(
+          fixedSize: Size(MediaQuery.of(context).size.width, 50),
         ),
         child: Row(
           mainAxisAlignment: MainAxisAlignment.spaceBetween,
