@@ -1,8 +1,8 @@
 abstract class ServicesUrls {
   //!Development
-  // static const String domain = 'https://widam.akwad.qa/api/method';
+  static const String domain = 'https://widam.akwad.qa/api/method';
   // !Production
-  static const String domain = 'https://app.widam.com.qa/api/method';
+  // static const String domain = 'https://app.widam.com.qa/api/method';
   static const String baseUrl = '$domain/widam_delivery';
   static const String googleMapsBaseUrl =
       'https://maps.googleapis.com/maps/api';
