@@ -15,7 +15,8 @@ class DealsScreen extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final topPadding = ref.watch(topPaddingProvider(MediaQuery.paddingOf(context).top));
+    final topPadding =
+        ref.watch(topPaddingProvider(MediaQuery.paddingOf(context).top));
     return Scaffold(
         body: Column(
       children: [
