@@ -39,7 +39,8 @@ class _ItemGroupScreenState extends ConsumerState<ItemGroupScreen> {
 
   @override
   Widget build(BuildContext context) {
-    final topPadding = ref.watch(topPaddingProvider(MediaQuery.paddingOf(context).top));
+    final topPadding =
+        ref.watch(topPaddingProvider(MediaQuery.paddingOf(context).top));
     return Scaffold(
       body: Column(
         children: [
