@@ -31,7 +31,8 @@ class _MubadaraItemsScreenState extends ConsumerState<MubadaraItemsScreen> {
   Widget build(BuildContext context) {
     final mubadaraItemsAsync =
         ref.watch(mubadaraItemsProvider(widget.mubadaraId));
-    final topPadding = ref.watch(topPaddingProvider(MediaQuery.paddingOf(context).top));
+    final topPadding =
+        ref.watch(topPaddingProvider(MediaQuery.paddingOf(context).top));
     return Scaffold(
       body: Column(
         children: [
