@@ -6,12 +6,11 @@ part 'banner.g.dart';
 
 @freezed
 class Banner with _$Banner implements BlockData {
-  const factory Banner({
-    required String bannerType,
-    required String? bannerLink,
-    required String bannerImage,
-    required String? mubadaraId
-  }) = _Banner;
+  const factory Banner(
+      {required String bannerType,
+      required String? bannerLink,
+      required String bannerImage,
+      required String? mubadaraId}) = _Banner;
 
   factory Banner.fromJson(Map<String, dynamic> json) => _$BannerFromJson(json);
 }
